refactor(codegen): extract std::io::file::* into stdlib/io_file.rs (Round 1.12)
6 fns lifted into IoFileStdlib<'ctx>: open_fallible, write_bytes_fallible,
seek_fallible, read_line, close, at_eof.

codegen.rs: 42235 -> 41865 (-370 lines).
